Move tests to node-http-incoming-socket.test.ts
node-http.test.ts has a pre-existing local failure (the http-proxy test)
that blocks the gate's pass check; a sibling file keeps the coverage in
test/js/node/http/ while running cleanly on its own.
